Created a result wrapper to allow returning status codes from handlers
before this change every handler that returned a OxiDish::http::Result would automatically have a status code of 200 (even if it was creating or modifying data). This means we were not meeting the spec for a REST api. Now that is fixed. We can also use this status code in testing to make sure nothing went wrong in our request without having to check the contents of the body.

the new custom result type (OxiDishResult) is not a proper implementation of std::Result, so it lacks the FromResidual trait that is required to use the ? operator in functions returning it. This is ok though because we can just match on anything we were using ? on, and manually call into() on any Error (because our custom Error derives From impls for any error that could be returned by any function in our program).

almost pushed code that didn't compile :sweat_smile: