Refactor model structs to better reflect their purpose, refactor get_recipe to get_recipe_by_id
This commit got ahead of me--it should be made up of smaller pieces, but
because git only allows line by line commits, a lot of the imports would
have gotten screwed up.

The goal here is to be specific about which structs refer to rows from
database queries, and which are used by clients of this library. This
commit:
- Explains why having a RecipeIngredientRow (which would normally be
id_recipe, id_ingredient_name, ingredient_amount) should be split up
into id_recipe and a separate struct IngredientEntry (which is used by
the client when creating new recipes)
- Renames a few structs to be more clear about which are modeled after
rows in the DB
- Refactors get_recipe to be get_recipe_by_id, again to more clearly
delineate which structs are used by clients (as parameters or return
values from _query_ functions), and which are internal models of the
database.

I'm sure there is a better way to organize these, as there isn't a clear
heirarchy. I moved NewRecipe out of models to be closer to the query
which depends on it, but it still depends on IngredientEntry, which is
in the models directory.
I also decided it makes sense for consumers of get_recipe_by_id to be
returned a _whole_ recipe, ingredients included, as opposed to just a
RecipeRow. This avoids the n+1 problem where a consumer would be forced
to query the database again (using something like
get_ingredients_for_recipe_id) to construct something actually useful to
a user. This necessitates creating another few structs (RecipeResponse
and IngredientResponse) that are specifically for clients to use.